requirements.txt
requests>=2.6.1
ruamel.yaml>=0.16.0,<0.19
rdflib>=4.2.2,<7.1
shellescape>=3.4.1,<3.9
schema-salad>=8.4.20230426093816,<9
prov==1.5.1
mypy-extensions
psutil>=5.6.6
importlib_resources>=1.4  # equivalent to Python 3.9
coloredlogs
pydot>=1.4.1
argcomplete>=1.12.0
pyparsing!=3.0.2  # breaks --print-dot (pydot) https://github.com/pyparsing/pyparsing/issues/319
cwl-utils>=0.22
